+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # v5.2.0 [Not released Yet]
 
+Added
+- Added deprecation warnings to `Connection::runDdl` and `Connection::runDdls` (#98)
+
 Changed
 - `Connection::waitForOperation` and `Connection::isDoneOperation` has been removed. (#99)